refactor: adjusted directories to adapt layered architecture
