+++ src/main/java/roomescape/controller/ExceptionController.java
@@ -1,8 +1,10 @@
-package roomescape;
+package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import roomescape.exception.NoParameterException;
+import roomescape.exception.NotFoundReservationException;
 
 @ControllerAdvice
 public class ExceptionController {
