+++ src/main/java/roomescape/exception/NoParameterException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class NoParameterException extends RuntimeException {
