+++ src/main/java/roomescape/exception/NotFoundTimeException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class NotFoundTimeException extends RuntimeException {
